Larger size to store radio results

+++ radio.h
@@ -29,7 +29,7 @@ typedef struct radio
 
 void radio_connect(radio_t *radio);
 void radio_disconnect(radio_t *radio);
-int radio_send_command(radio_t *radio, uint8_t cmd, int16_t subcmd, int16_t subsubcmd, uint8_t *data, int len, uint32_t *return_value, uint8_t reverse_value);
+int radio_send_command(radio_t *radio, uint8_t cmd, int16_t subcmd, int16_t subsubcmd, uint8_t *data, int len, uint64_t *return_value, uint8_t reverse_value);
 int radio_set_satellite_mode(radio_t *radio, int sat_mode);
